% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wca2d.R
\name{wca2d}
\alias{wca2d}
\title{Runs wca2d for a set number of time steps}
\usage{
wca2d(dem, ppt, wse, deltat, deltatu, currt, nullvall = 1e+06, mannN = 0.05,
  cellem = 50)
}
\arguments{
\item{dem}{A matrix with elevation values in each cell (m)}

\item{ppt}{A matrix with precipitation values in each cell (mm)}

\item{wse}{A matrix with the water surface elevation values from the previous iteration(m)}

\item{deltat}{The iteration time step (s)}

\item{deltatu}{The update time step (s)}

\item{currt}{The current simulation time (from 0, s)}

\item{mannN}{Manning's coefficent}

\item{cellem}{The cell dimensions (m). Note that this currently assumes a constant and regular grid}

\item{nullval}{Value for any grid cells to be ignored}
}
\value{
\code{simcf}: a list of output from the Fortran routine
}
\description{
Function to run the weighted cellular automata hydrological model for multiple time steps.
TBD.
}
\examples{
wca2d_1t(dem, ppt, wse, 1, 60, 0)
}
